Update APD_traits.csv
There were 154 nutrient content traits where there was never a "non-encoded" description. For the new traits table we'll put on Zenodo, I think it will be good for all traits to have a "plain English" description. The descriptions added are identical to those in the encoded descriptions, except without the references to published classes. This was mostly done with
```
  mutate(
    description = ifelse(is.na(description), description_encoded, description),
    description = stringr::str_replace_all(description, "[:space:]\\[[:alnum:]+[:punct:][:alnum:]+\\]", "")
    )
```

touching up a few encodings that were missed.