Minor fixes (#32)
* fix errors in GIFT trait names

* fix errors in TRY trait names

There were 63 instances of TRY names/trait codes that did not match the current list of TRY names/trait codes. 3 of those were instances where APD had a mismatched name-code for a TRY match and the remainder were where TRY had slightly edited their trait names since our previous matches. We have continued to replace ";" with "," in TRY trait names, because ";" are a delimiter used in APD.

Also  found a few additional matches to TRY traits to add. This is not a comprehensive review of additional trait matches that might exist, but simply adding a few that were apparent.

closes https://github.com/traitecoevo/APD/issues/29
closes https://github.com/traitecoevo/APD/issues/28