functions to convert between csv and yml formats
A pair of functions to convert the APD_traits_input.csv into a yml file (in the long run, the core storage format) and a csv for editing.

Testing by converting back and forth several times and no changes.

For now, will retain the csv, but eventually the plan will be to have the yml the only version that has commits and then the yml will be converted to csv for releases.

Confirmed that the only changes to ttl, nt, etc files are the expected supression of scientific notation and the recording of the two out-of-order traits.